+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Compatibility:
 * Implement `String#undump` (#2131, @kustosz)
 * `Errno` constants with the same `errno` number are now the same class.
 * Implement `Enumerable#tally` and `Enumerable#filter_map` (#2144 and #2152, @LillianZ).
+* Implement `Range#minmax`.
 
 Performance: